refactor: standardize test scripts and improve documentation
