+++ package.json
@@ -51,9 +51,9 @@
     "lint:fix": "eslint . --fix && npm run lint:fix -w web",
     "format": "prettier --check .",
     "format:fix": "prettier --write .",
-    "test:cli": "tsx --test tests/e2e.test.ts",
+    "test:cli": "tsx --test tests/cli.spec.ts",
     "test:unit": "tsx --test src/**/*.test.ts && npm run test:unit -w web",
-    "test:web": "npm run test -w web",
+    "test:web": "npm run test:web -w web",
     "test:storybook": "npm run test:storybook -w web",
     "test:visual": "npm run test:visual -w web",
     "test:visual:update": "npm run test:visual:update -w web",
